Rename "variable.dart" to "bool_variable.dart".
