+++ patterns/interpreter/conceptual/variable/bool_variable.dart
